move marshal functionality to json
