+++ components.go
@@ -2,7 +2,6 @@ package discordgo
 
 import (
 	"encoding/json"
-	"fmt"
 )
 
 // ComponentType is type of component.
@@ -22,78 +21,9 @@ type MessageComponent interface {
 	Type() ComponentType
 }
 
-type unmarshalableMessageComponent struct {
-	MessageComponent
-}
-
-// UnmarshalJSON is a helper function to unmarshal MessageComponent object.
-func (umc *unmarshalableMessageComponent) UnmarshalJSON(src []byte) error {
-	var v struct {
-		Type ComponentType `json:"type"`
-	}
-	err := json.Unmarshal(src, &v)
-	if err != nil {
-		return err
-	}
-
-	switch v.Type {
-	case ActionsRowComponent:
-		umc.MessageComponent = &ActionsRow{}
-	case ButtonComponent:
-		umc.MessageComponent = &Button{}
-	case SelectMenuComponent:
-		umc.MessageComponent = &SelectMenu{}
-	case TextInputComponent:
-		umc.MessageComponent = &TextInput{}
-	default:
-		return fmt.Errorf("unknown component type: %d", v.Type)
-	}
-	return json.Unmarshal(src, umc.MessageComponent)
-}
-
-// MessageComponentFromJSON is a helper function for unmarshaling message components
-func MessageComponentFromJSON(b []byte) (MessageComponent, error) {
-	var u unmarshalableMessageComponent
-	err := u.UnmarshalJSON(b)
-	if err != nil {
-		return nil, fmt.Errorf("failed to unmarshal into MessageComponent: %w", err)
-	}
-	return u.MessageComponent, nil
-}
-
 // ActionsRow is a container for components within one row.
 type ActionsRow struct {
 	Components []MessageComponent `json:"components"`
-}
-
-// MarshalJSON is a method for marshaling ActionsRow to a JSON object.
-func (r ActionsRow) MarshalJSON() ([]byte, error) {
-	type actionsRow ActionsRow
-
-	return json.Marshal(struct {
-		actionsRow
-		Type ComponentType `json:"type"`
-	}{
-		actionsRow: actionsRow(r),
-		Type:       r.Type(),
-	})
-}
-
-// UnmarshalJSON is a helper function to unmarshal Actions Row.
-func (r *ActionsRow) UnmarshalJSON(data []byte) error {
-	var v struct {
-		RawComponents []unmarshalableMessageComponent `json:"components"`
-	}
-	err := json.Unmarshal(data, &v)
-	if err != nil {
-		return err
-	}
-	r.Components = make([]MessageComponent, len(v.RawComponents))
-	for i, v := range v.RawComponents {
-		r.Components[i] = v.MessageComponent
-	}
-
-	return err
 }
 
 // Type is a method to get the type of a component.
@@ -137,23 +67,6 @@ type Button struct {
 	CustomID string `json:"custom_id,omitempty"`
 }
 
-// MarshalJSON is a method for marshaling Button to a JSON object.
-func (b Button) MarshalJSON() ([]byte, error) {
-	type button Button
-
-	if b.Style == 0 {
-		b.Style = PrimaryButton
-	}
-
-	return json.Marshal(struct {
-		button
-		Type ComponentType `json:"type"`
-	}{
-		button: button(b),
-		Type:   b.Type(),
-	})
-}
-
 // Type is a method to get the type of a component.
 func (Button) Type() ComponentType {
 	return ButtonComponent
@@ -188,19 +101,6 @@ func (SelectMenu) Type() ComponentType {
 	return SelectMenuComponent
 }
 
-// MarshalJSON is a method for marshaling SelectMenu to a JSON object.
-func (m SelectMenu) MarshalJSON() ([]byte, error) {
-	type selectMenu SelectMenu
-
-	return json.Marshal(struct {
-		selectMenu
-		Type ComponentType `json:"type"`
-	}{
-		selectMenu: selectMenu(m),
-		Type:       m.Type(),
-	})
-}
-
 // TextInput represents text input component.
 type TextInput struct {
 	CustomID    string         `json:"custom_id"`
@@ -216,19 +116,6 @@ type TextInput struct {
 // Type is a method to get the type of a component.
 func (TextInput) Type() ComponentType {
 	return TextInputComponent
-}
-
-// MarshalJSON is a method for marshaling TextInput to a JSON object.
-func (m TextInput) MarshalJSON() ([]byte, error) {
-	type inputText TextInput
-
-	return json.Marshal(struct {
-		inputText
-		Type ComponentType `json:"type"`
-	}{
-		inputText: inputText(m),
-		Type:      m.Type(),
-	})
 }
 
 // TextInputStyle is style of text in TextInput component.
