+++ json_test.go
@@ -1,0 +1,13 @@
+package discordgo
+
+import (
+	"errors"
+	"testing"
+)
+
+func Test_unmarshal(t *testing.T) {
+	err := unmarshal([]byte{}, &struct{}{})
+	if !errors.Is(err, ErrJSONUnmarshal) {
+		t.Errorf("Unexpected error type: %T", err)
+	}
+}
