+++ restapi_test.go
@@ -1,7 +1,6 @@
 package discordgo
 
 import (
-	"errors"
 	"testing"
 )
 
@@ -228,10 +227,3 @@ func TestGuildPrune(t *testing.T) {
 	}
 }
 */
-
-func Test_unmarshal(t *testing.T) {
-	err := unmarshal([]byte{}, &struct{}{})
-	if !errors.Is(err, ErrJSONUnmarshal) {
-		t.Errorf("Unexpected error type: %T", err)
-	}
-}
